pyproject.toml
[project]
name = "fis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "click==8.1.8",
    "ipykernel>=6.29.5",
    "matplotlib==3.10.0",
    "networkx==3.4.2",
    "numpy>=1.16.2,<2.0.0",
    "pytest==8.3.4",
    "python-dotenv==1.0.1",
    "scikit-learn==1.6.0",
    "seaborn==0.13.2",
    "torch==2.5.1",
    "torchinfo==1.8.0",
    "torchvision==0.20.1",
    "tqdm>=4.67.1",
    "wandb==0.19.1",
]

[project.scripts]
fis = "fis:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
